requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
numpy~=2.3.1
scikit-learn~=1.7.0
google-genai~=1.24.0
python-dotenv~=1.1.1
apiflask~=2.4.0
pywebview[qt]~=5.4
Flask~=3.1.1
feedparser~=6.0.11
lxml~=6.0.0
pyshortcuts~=1.9.5
Flask-SQLAlchemy~=3.1.1
Flask-JWT-Extended~=4.7.1
Flask-Migrate~=4.1.0